# doc/api.yaml
openapi: 3.0.3
info:
  title: wasa-text api
  description: >
    This is the API documentation for the wasa-text project.
    The project is a chat application that allows users to chat with each other.
    The API allows users to log in, create groups, send messages, and more.
  version: "1"
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /users:
    get:
      tags: ["user"]
      security:
        - bearerAuth: []
      summary: Get user info
      description: Get user info from its id
      operationId: getUsers
      responses:
        "200":
          description: Return all users info
          content:
            application/json:
              schema:
                description: List of users
                type: array
                items:
                  $ref: "#/components/schemas/User"
                default: []
                minItems: 0
                example: [{"id": 1,"name": "mariOne","photo": "null"},{"id": 2,"name": "rino","photo": "null"},{"id": 3,"name": "gianna","photo": "null"}]
        "401":
          $ref: "#/components/responses/401_Unauthorized"
        "404":
          $ref: "#/components/responses/404_NotFound"
  /users/{user_id}:
    get:
      tags: ["user"]
      security:
        - bearerAuth: []
      summary: Get user info
      description: Get user info from its id
      operationId: getUser
      parameters:
        - name: user_id
          in: path
          required: true
          description: The user identifier
          schema:
            $ref: "#/components/schemas/userId"
      responses:
        "200":
          description: Return user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/401_Unauthorized"
        "404":
          $ref: "#/components/responses/404_NotFound"
...
  /conversations/{conversation_id}/messages:
    parameters:
      - name: conversation_id
        in: path
        required: true
        description: Conversation identifier
        schema:
          # description: The id of the conversation to get the messages
          $ref: "#/components/schemas/conversationId"
    get:
      tags: ["message"]
      security:
        - bearerAuth: []
      summary: Get all messages of conversation
      description: Get all messages of conversation by its id
      operationId: getMessages
      responses:
        "200":
          description: Return all messages of the conversation
          content:
            application/json:
              schema:
                description: List of messages
                type: array
                minItems: 0
                default: []
                items:
                  $ref: "#/components/schemas/Message"
                example: [{"id": 3,"text": "Brella","photo": "null","author": 2,"recipient": 2,"forward": 0,"timestamp": "2025-01-22T17:22:15Z"},{"id": 4,"text": "Ma dove vai?","photo": "null","author": 3,"recipient": 2,"forward": 0,"timestamp": "2025-01-22T17:23:18Z"}]
        "404":
          $ref: "#/components/responses/404_NotFound"
    post:
      tags: ["message"]
      security:
        - bearerAuth: []
      summary: Send a message
      description: Send a message to a conversation
      operationId: sendMessage
      requestBody:
        description: Message details
        required: true
        content:
          multipart/form-data:
            schema:
              description: Message details for sending a message
              type: object
              properties:
                text:
                  $ref: "#/components/schemas/messageText"
                photo:
                  $ref: "#/components/schemas/messagePhoto"
                reply:
                  $ref: "#/components/schemas/messageId"

      responses:
        "204":
          $ref: "#/components/responses/204_NoContent"
        "404":
          $ref: "#/components/responses/404_NotFound"
  /conversations/{conversation_id}/messages/{message_id}:
    delete:
      tags: ["message"]
      security:
        - bearerAuth: []
      summary: Delete a message
      description: Delete a Message by its Id
      operationId: deleteMessage
      parameters:
        - name: message_id
          in: path
          required: true
          description: The message identifier
          schema:
            $ref: "#/components/schemas/messageId"
            # description: The id of the message to delete
        - name: conversation_id
          in: path
          required: true
          description: The conversation identifier
          schema:
            $ref: "#/components/schemas/conversationId"
            # description: The id of the conversation of the message
      responses:
        "204":
          $ref: "#/components/responses/204_NoContent"
        "401":
          $ref: "#/components/responses/401_Unauthorized"
        "404":
          $ref: "#/components/responses/404_NotFound"
  /conversations/{conversation_id}/messages/{message_id}/forward:
    post:
      tags: ["message"]
      security:
        - bearerAuth: []
      summary: Forward a message
      description: Forward a message to another conversation
      operationId: forwardMessage
      parameters:
        - name: message_id
          in: path
          required: true
          description: The message identifier
          schema:
            $ref: "#/components/schemas/messageId"
            # description: The id of the message to forward
        - name: conversation_id
          in: path
          required: true
          description: The conversation identifier
          schema:
            $ref: "#/components/schemas/conversationId"
            # description: The id of the conversation of the message
      responses:
        "204":
          $ref: "#/components/responses/204_NoContent"
        "401":
          $ref: "#/components/responses/401_Unauthorized"
        "404":
          $ref: "#/components/responses/404_NotFound"
...
components:
  schemas:
    userId:
      description: user id properties
      type: integer
      example: 1
      minimum: 1
      maximum: 999
    userName:
      description: user name properties
      type: string
      example: Mario
      pattern: "^[a-zA-Z0-9]+$"
      minLength: 3
      maxLength: 30
    userPhoto:
      description: user photo properties
      type: string
      format: binary
    User:
      description: Schema representing a user object
      type: object
      properties:
        id:
          $ref: "#/components/schemas/userId"
        name:
          $ref: "#/components/schemas/userName"
        photo:
          $ref: "#/components/schemas/userPhoto"
      
      example:
        id: 1
        name: Mario
        photo: "null"
    conversationId:
      description: conversation id properties
      type: integer
      example: 1
      minimum: 1
      maximum: 999
...
    messageId:
      description: message id
      type: integer
      example: 1
      minimum: 1
      maximum: 999
    messageText:
      description: message text properties
      type: string
      example: "Ciao"
      pattern: "^.+$" # tutto tranne stringa vuota
      minLength: 1
      maxLength: 200
    messagePhoto:
      description: message photo properties
      type: string
      format: binary
      minLength: 1
      maxLength: 10485760 # 10 MB
    messageTimestamp:
      description: date time of the message creation
      type: string
      format: date-time
      example: "2024-12-26T14:16:10Z"
    Message:
      type: object
      description: Schema representing a message object
      properties:
        text:
          $ref: "#/components/schemas/messageText"
        photo:
          $ref: "#/components/schemas/messagePhoto"
        author:
          # description: User Id of the message sender
          $ref: "#/components/schemas/userId"
        recipient:
          # description: Conversation Id of the recipient conversation
          $ref: "#/components/schemas/conversationId"
        forward:
          # description: If forwarded, the user id of the original author
          $ref: "#/components/schemas/userId"
        reply:
          # description: If reply, the message id of the replied message
          $ref: "#/components/schemas/messageId"
        timestamp:
          $ref: "#/components/schemas/messageTimestamp"
        reactions:
          type: array
          description: Schema representing a reaction object
          items:
            $ref: "#/components/schemas/Reaction"
          default: []
          minItems: 0
      example:
        text: "Ciao"
        photo: "null"
        author: 1
        recipient: 1
        forward: 0
        reply: 0
        timestamp: "2024-12-26T14:16:10Z"
        reactions: []
    reactionText:
      description: reaction text properties
      type: string
      example: "🗿"
      pattern: "^(😂|🗿|😐|👍|❤️|🔥|🎉|😢|😡)$"
      minLength: 1
      maxLength: 1
    Reaction:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/userId"
        reaction:
          $ref: "#/components/schemas/reactionText"
      example:
        user: 1
        reaction: "🗿"
  responses:
...
    204_NoContent:
      description: No content, operation successful
    401_Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            description: Unauthorized response
            type: object
            properties:
              message:
                description: Unauthorized message
                type: string
                example: "Unauthorized"
                minLength: 0
                maxLength: 200
                pattern: "^.*?$"
...
    404_NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            description: Not Found response
            type: object
            properties:
              message:
                description: Not Found message
                type: string
                example: "Not Found"
                minLength: 0
                maxLength: 200
                pattern: "^.*?$"
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: apiKey
      in: header
      name: authorization